implemented system to drop different types of chips (calculating least possible total chips dropped) when enemy dies

+++ Assets/Scripts/Enemy/Universal/EnemyStats.cs
@@ -10,7 +10,7 @@ public class EnemyStats : ScriptableObject
     [Header("Base Stats")]
     public float MaxHealth;
     public float MoveSpeed;
-    public float ChipsDropped;
+    public int ChipsDropped;
 
     [Header("Shooting Stats")]
     public float BulletSpeed;
